Associate incoming texts to all clients associated with that phone number
This is a band-aid fix to a larger issue with duplicate clients. For now, we will just associate messages with ALL clients with that phone number. This should really probably be done in a background job, but I don't want to make too many changes all at once.
