+++ app/services/incoming_text_message_service.rb
@@ -1,0 +1,50 @@
+class IncomingTextMessageService
+  attr_accessor :params
+
+  def self.process(params)
+    phone_number = PhoneParser.normalize(params["From"])
+
+    clients = Client.joins(:intake).where(intakes: { phone_number: phone_number} ).or(Client.joins(:intake).where(intakes: { sms_phone_number: phone_number}))
+
+    unless clients.exists?
+      clients = [Client.create!(
+        intake: Intake.create!(
+          phone_number: phone_number,
+          sms_phone_number: phone_number,
+
+          visitor_id: SecureRandom.hex(26),
+          sms_notification_opt_in: "yes",
+        ),
+        vita_partner: VitaPartner.client_support_org,
+      )]
+    end
+
+    # process attachments once
+    attachments = TwilioService.new(params).parse_attachments
+
+    clients.map do |client|
+      documents = attachments.map do |attachment|
+        Document.new(
+          client: client,
+          document_type: DocumentTypes::TextMessageAttachment.key,
+          upload: {
+              io: StringIO.new(attachment[:body]),
+              filename: attachment[:filename],
+              content_type: attachment[:content_type],
+              identify: false
+          }
+        )
+      end
+
+      contact_record = IncomingTextMessage.create!(
+        body: params["Body"],
+        received_at: DateTime.now,
+        from_phone_number: phone_number,
+        client: client,
+        documents: documents
+      )
+
+      ClientChannel.broadcast_contact_record(contact_record)
+    end
+  end
+end
