+++ spec/controllers/twilio_webhooks_controller_spec.rb
@@ -27,7 +27,7 @@ RSpec.describe TwilioWebhooksController do
       }
     end
 
-    describe "#create" do
+    describe "#create_incoming_text_message" do
       context "with an invalid request" do
         before do
           allow(TwilioService).to receive(:valid_request?).and_return false
@@ -41,115 +41,15 @@ RSpec.describe TwilioWebhooksController do
       end
 
       context "with a valid request" do
-        let(:current_time) { DateTime.new(2020, 9, 6) }
         before do
           allow(TwilioService).to receive(:valid_request?).and_return true
-          allow(DateTime).to receive(:now).and_return current_time
+          allow(IncomingTextMessageService).to receive(:process)
         end
 
-        context "with a matching intake phone number" do
-          let(:client) { create :client }
-          let!(:intake) { create(:intake, client: client, phone_number: "+15005550006") }
+        it "processes the text in the IncomingTextMessageService" do
+          post :create_incoming_text_message, params: incoming_message_params
 
-          it "creates a new IncomingTextMessage linked to the client the right data" do
-            expect do
-              post :create_incoming_text_message, params: incoming_message_params
-            end.to change(IncomingTextMessage, :count).by 1
-
-            expect(response).to be_ok
-            message = IncomingTextMessage.last
-            expect(message.body).to eq "Hello, it me"
-            expect(message.from_phone_number).to eq "+15005550006"
-            expect(message.received_at).to eq current_time
-            expect(message.client).to eq client
-          end
-        end
-
-        context "with a matching client sms_phone_number" do
-          before do
-            allow(ClientChannel).to receive(:broadcast_contact_record)
-          end
-          let(:intake) { create(:intake, sms_phone_number: "+15005550006")}
-          let!(:client) { create :client, intake: intake }
-
-          it "creates a new IncomingTextMessage linked to the client the right data" do
-            expect do
-              post :create_incoming_text_message, params: incoming_message_params
-            end.to change(IncomingTextMessage, :count).by 1
-
-            message = IncomingTextMessage.last
-            expect(message.client).to eq client
-          end
-
-          it "sends a real-time update to anyone on this client's page" do
-            post :create_incoming_text_message, params: incoming_message_params
-            expect(ClientChannel).to have_received(:broadcast_contact_record).with(IncomingTextMessage.last)
-          end
-        end
-
-        context "with multiple matching clients" do
-          # We have not discussed the best way to handle this scenario
-          # This spec is intended to document existing behavior more than
-          # prescribe the correct way to handle this.
-          let(:intake1) { create :intake, phone_number: "+15005550006" }
-          let(:intake2) { create :intake, sms_phone_number: "+15005550006" }
-          let!(:client1) { create :client, intake: intake1 }
-          let!(:client2) { create :client, intake: intake2 }
-
-          it "creates a new IncomingTextMessage linked to the first client" do
-            expect do
-              post :create_incoming_text_message, params: incoming_message_params
-            end.to change(IncomingTextMessage, :count).by 1
-
-            message = IncomingTextMessage.last
-            expect(message.client).to eq client1
-          end
-        end
-
-        context "without a matching client" do
-          it "creates a new incoming text message attached to a new client" do
-            expect do
-              post :create_incoming_text_message, params: incoming_message_params
-            end.to change(IncomingTextMessage, :count).by(1).and change(Client, :count).by(1)
-
-            message = IncomingTextMessage.last
-            expect(message.body).to eq "Hello, it me"
-            expect(message.from_phone_number).to eq "+15005550006"
-            expect(message.received_at).to eq current_time
-            client = Client.last
-            expect(message.client).to eq client
-            expect(client.intake.phone_number).to eq "+15005550006"
-            expect(client.intake.sms_phone_number).to eq "+15005550006"
-            expect(client.intake.sms_notification_opt_in).to eq("yes")
-            expect(client.vita_partner).to eq VitaPartner.client_support_org
-          end
-        end
-
-        context "with an attachment" do
-          let!(:client) { create :client }
-          let!(:intake) { create :intake, client: client, sms_phone_number: "+15005550006" }
-          let(:body) { "" }
-          let(:parsed_attachments) do
-            [{content_type: "image/jpeg", filename: "some-type-of-image.jpg", body: "image file contents"}]
-          end
-
-          before do
-            allow(ClientChannel).to receive(:broadcast_contact_record)
-            #TODO: use fake instead
-            allow_any_instance_of(TwilioService).to receive(:parse_attachments).and_return(parsed_attachments)
-          end
-
-          it "creates a new IncomingTextMessage linked to the client the right data" do
-            post :create_incoming_text_message, params: incoming_message_params
-
-            documents = client.documents
-
-            expect(documents.count).to eq(1)
-            expect(documents.all.pluck(:document_type).uniq).to eq([DocumentTypes::TextMessageAttachment.key])
-            expect(documents.first.contact_record).to eq IncomingTextMessage.last
-            expect(documents.first.upload.blob.download).to eq("image file contents")
-            expect(documents.first.upload.blob.content_type).to eq("image/jpeg")
-          end
+          expect(IncomingTextMessageService).to have_received(:process).with(hash_including(incoming_message_params))
         end
       end
     end
@@ -164,7 +64,7 @@ RSpec.describe TwilioWebhooksController do
       end
 
       it "returns a 403 status code" do
-        post :update_outgoing_text_message, params: {id: existing_message.id}
+        post :update_outgoing_text_message, params: { id: existing_message.id }
 
         expect(response.status).to eq 403
       end
